+++ backend/srcs/src/auth/auth.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, UnauthorizedException } from '@nestjs/common';
 import { JwtService } from '@nestjs/jwt';
 import { JwtPayload } from './strategy';
 import { User } from '@prisma/client';
@@ -15,5 +15,21 @@ export class AuthService {
     return {
       accessToken: this.jwtService.sign(payload),
     };
+  }
+
+  async isAuthorized(request: any ){
+    const { token } = request.handshake.auth;
+    if (!token) {
+        throw new UnauthorizedException();
+    }
+    try {
+        const payload = await this.jwtService.verifyAsync(
+            token,
+            {   secret: process.env.JWT_SECRET    }
+        );
+        return payload
+    } catch {
+        throw new UnauthorizedException();
+    }
   }
 }