+++ backend/srcs/src/auth/guard/index.ts
@@ -1,2 +1,3 @@
 export * from './fortytwo.guard';
 export * from './jwt.guard';
+export * from './ws.guard';