+++ backend/srcs/src/chat/dto/index.ts
@@ -1,2 +1,3 @@
 export * from './createChatMessage.dto';
 export * from './updateChatMessage.dto';
+export * from './createChatMessageWs.dto'