+++ backend/srcs/src/main.ts
@@ -9,7 +9,12 @@ import * as session from 'express-session';
 import * as passport from 'passport';
 
 async function bootstrap() {
-  const app = await NestFactory.create(AppModule);
+  const app = await NestFactory.create(AppModule, { cors: true });
+  // const app = await NestFactory.create(AppModule);
+  app.enableCors({
+    origin: '*',
+    credentials: true,
+  });
   app.useGlobalPipes(
     new ValidationPipe({
       whitelist: true,
@@ -50,6 +55,6 @@ async function bootstrap() {
     },
   });
 
-  await app.listen(3000);
+  await app.listen(process.env.API_PORT);
 }
 bootstrap();