+++ frontend/srcs/src/app/app-routing.module.ts
@@ -1,23 +1,25 @@
 import { NgModule } from '@angular/core';
-import { RouterModule, Routes } from '@angular/router';
+import { PreloadAllModules, RouterModule, Routes } from '@angular/router';
 import { HomeComponent } from './modules/home/home.component';
-import { AppComponent } from './app.component';
 import { GameComponent } from './modules/game/game.component';
 import { HistoryComponent } from './modules/history/history.component';
 import { ChatComponent } from './modules/chat/chat.component';
 import { ProfileComponent } from './modules/profile/profile.component';
+import {  LoginGuard } from './guards';
 
 const routes: Routes = [
+  
   {path:"", redirectTo:"", pathMatch:"full"},
-  {path:"home", component:HomeComponent},
-  {path:"game", component:GameComponent},
-  {path:"history", component:HistoryComponent},
-  {path:"chat", component:ChatComponent},
-  {path:"profile", component:ProfileComponent}
+  {path:"", component:HomeComponent, canActivate: [LoginGuard]},
+  {path:"home", component:HomeComponent, canActivate: [LoginGuard]},
+  {path:"game", component:GameComponent,canActivate: [LoginGuard]},
+  {path:"history", component:HistoryComponent, canActivate: [LoginGuard]},
+  {path:"chat", component:ChatComponent, canActivate: [LoginGuard]},
+  {path:"profile", component:ProfileComponent, canActivate: [LoginGuard]}
 ];
 
 @NgModule({
-  imports: [RouterModule.forRoot(routes)],
+  imports: [RouterModule.forRoot(routes, {preloadingStrategy: PreloadAllModules})],
   exports: [RouterModule]
 })
 export class AppRoutingModule { }