+++ frontend/srcs/src/app/models/core/auth.model.ts
@@ -1,0 +1,15 @@
+export namespace AuthModel {
+  export interface UserTokenData {
+    iat:          number;
+    exp:          number;
+    username:     string;
+    sub:          string;
+  }
+
+  export const userTokenData: UserTokenData = {
+    sub: '',
+    username: '',
+    iat: 0,
+    exp: 0
+  }
+}